Prove that the stated specification admits simultaneously eliminating candidates.
Major progress on this proof.  The main content of this proof is that
a collection of candidates may be eliminated simultaneously if the combined
first-choice votes for all those candidates in a round is less than
the individual counts of all other candidates.  In this situation,
none of the candididates to be eliminated could possibly win because
even all of their consolidated votes could not surpass any other
remaining candidates, no matter how those votes are apportioned.